code: rename test files for react-scripts v3
