+++ src/components/explore-space-detail-utilization-card/index.test.js
