+++ src/helpers/auto-refresh-hoc/index.test.js
