+++ src/helpers/filter-collection/index.test.js
