+++ src/helpers/format-capacity-percentage/index.test.js
