+++ src/helpers/format-percentage/index.test.js
