+++ src/helpers/format-tag-name/index.test.js
