+++ src/helpers/space-hierarchy-formatter-deprecated/index.test.js
