+++ src/helpers/space-hierarchy-formatter/index.test.js
