+++ src/helpers/space-hierarchy-searcher/index.test.js
