+++ src/helpers/string-to-boolean/index.test.js
