+++ src/reducers/account-registration/index.test.js
