+++ src/reducers/active-modal/index.test.js
