+++ src/reducers/sensors/index.test.js
