+++ src/reducers/spaces/index.test.js
